Add phase-based camera routes with random jumps, indefinite duration, and ring wrap-copy scroll demos
- engine: widen frame_index/run_frames to u32 (indefinite runs)
- tile_scroll: surface prefetch size as template param; DESC blits for overlapping shifts; max_blit_jobs 64
- 101/102: RouteCamera with horizontal/vertical/diagonal/circular/sinusoidal phases + pauses, jump mode (2-15px) after ~1min re-randomized every 10s; 102 re-randomizes tile patterns every 10s
- 103/104: Lionheart-style ring scroll on 352x288 (2-tile margin), shifting the buffer 16px and reloading edge tiles on wrap; 103 EHB, 104 dual 3+3
- add eng/scene/route_camera.hpp

+++ engine/include/eng/graphics/driver.hpp
@@ -35,7 +35,7 @@ enum class GraphicsDriverId : u8 {
 /// iran alimentando desde blitter, copper y sprites para que las demos puedan fallar
 /// antes de saturar el hardware.
 struct FrameStats {
-	u16 frame_index = 0;
+	u32 frame_index = 0;
 	u16 blit_jobs = 0;
 	u16 copper_patches = 0;
 	u16 hardware_sprites = 0;
